// brute-api/src/tokens/token-type.enum.ts
export enum TokenType {
  ACTIVATION = "ACTIVATION",
  PASSWORD_RESET = "PASSWORD_RESET",
  PRE_AUTH = "PRE_AUTH",
  REFRESH = "REFRESH",
  SESSION = "SESSION",
  PASSWORD_CHANGE = "PASSWORD_CHANGE",
  EMAIL_CHANGE = "EMAIL_CHANGE",
  MFA_RESET = "MFA_RESET",
  PHONE_CHANGE = "PHONE_CHANGE",
  FRAUD_FLAG = "FRAUD_FLAG",
  REACTIVATION = "REACTIVATION",
}